refactor: move display-order functions from conversation-store to conversation-crud
