+++ assistant/src/memory/conversation-crud.ts
@@ -11,7 +11,8 @@ import { getLogger } from "../util/logger.js";
 import { createRowMapper } from "../util/row-mapper.js";
 import { deleteOrphanAttachments } from "./attachments-store.js";
 import { projectAssistantMessage } from "./conversation-attention-store.js";
-import { getDb, rawExec, rawGet } from "./db.js";
+import { ensureDisplayOrderMigration } from "./conversation-display-order-migration.js";
+import { getDb, rawExec, rawGet, rawRun } from "./db.js";
 import { indexMessageNow } from "./indexer.js";
 import {
   channelInboundEvents,
@@ -892,4 +893,55 @@ export function getConversationRecentProvenanceTrustClass(
   } catch {
     return undefined;
   }
+}
+
+// ---------------------------------------------------------------------------
+// CRUD functions for display_order and is_pinned
+// ---------------------------------------------------------------------------
+
+export function batchSetDisplayOrders(
+  updates: Array<{
+    id: string;
+    displayOrder: number | null;
+    isPinned: boolean;
+  }>,
+): void {
+  ensureDisplayOrderMigration();
+  rawExec("BEGIN");
+  try {
+    for (const update of updates) {
+      rawRun(
+        "UPDATE conversations SET display_order = ?, is_pinned = ? WHERE id = ?",
+        update.displayOrder,
+        update.isPinned ? 1 : 0,
+        update.id,
+      );
+    }
+    rawExec("COMMIT");
+  } catch (err) {
+    rawExec("ROLLBACK");
+    throw err;
+  }
+}
+
+export function getDisplayMetaForConversations(
+  conversationIds: string[],
+): Map<string, { displayOrder: number | null; isPinned: boolean }> {
+  ensureDisplayOrderMigration();
+  const result = new Map<
+    string,
+    { displayOrder: number | null; isPinned: boolean }
+  >();
+  if (conversationIds.length === 0) return result;
+  for (const id of conversationIds) {
+    const row = rawGet<{
+      display_order: number | null;
+      is_pinned: number | null;
+    }>("SELECT display_order, is_pinned FROM conversations WHERE id = ?", id);
+    result.set(id, {
+      displayOrder: row?.display_order ?? null,
+      isPinned: (row?.is_pinned ?? 0) === 1,
+    });
+  }
+  return result;
 }
